Move BitmapValue to package plugin-common
