+++ fe/plugin-common/src/main/java/com/starrocks/common/io/Codec.java
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package com.starrocks.common;
+package com.starrocks.common.io;
 
 import java.io.DataInput;
 import java.io.DataOutput;
