+++ fe/spark-dpp/src/main/java/com/starrocks/load/loadv2/dpp/StarRocksKryoRegistrator.java
@@ -18,6 +18,8 @@
 package com.starrocks.load.loadv2.dpp;
 
 import com.esotericsoftware.kryo.Kryo;
+import com.starrocks.common.io.BitmapValue;
+import com.starrocks.common.io.Roaring64Map;
 import org.apache.spark.serializer.KryoRegistrator;
 
 /**
